% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_journal.R
\name{acct.at.depth}
\alias{acct.at.depth}
\title{Returns account stringth to a given depth}
\usage{
acct.at.depth(acct, depth = NULL)
}
\arguments{
\item{string}{input account name}

\item{numeric}{desired account depth (defaults to null, which will return the original account)}
}
\value{
string shortened account name
}
\description{
Returns account stringth to a given depth
}
