% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_dataframe_journal.R
\name{as.data.frame.journal}
\alias{as.data.frame.journal}
\title{Converts journal to a dataframe of entries}
\usage{
\method{as.data.frame}{journal}(journal)
}
\arguments{
\item{journal}{input}
}
\value{
dataframe giving rows corresponding to each element in input journal
}
\description{
Converts journal to a dataframe of entries
}
\seealso{
\code{\link{read.journal}}
}
