% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_export.R
\name{balance_export}
\alias{balance_export}
\title{Title}
\usage{
balance_export(path_to_export, export_type, ledger_turnover)
}
\arguments{
\item{ledger_turnover}{}
}
\description{
Title
}
\keyword{internal}
