% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ledger.R
\name{check_ledger}
\alias{check_ledger}
\title{Tests for ledger}
\usage{
check_ledger(path_to_ledger)
}
\arguments{
\item{path_to_ledger}{Path to short ledger.}
}
\description{
Tests for ledger
}
\keyword{internal}
