% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ledger_dir.R
\name{check_ledger_dir}
\alias{check_ledger_dir}
\title{Test for ledger dir}
\usage{
check_ledger_dir(path_to_ledgerdir)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files}
}
\description{
Test for ledger dir
}
\keyword{internal}
