% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_routine.R
\name{check_routine}
\alias{check_routine}
\title{Title}
\usage{
check_routine(
  path_to_ledgerdir = F,
  path_to_export = F,
  export_type = F,
  ledger_path = F,
  ledger = F,
  path_to_maptab = F
)
}
\arguments{
\item{path_to_maptab}{}
}
\description{
Title
}
\keyword{internal}
