% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_export.R
\name{clean_export}
\alias{clean_export}
\title{Title}
\usage{
clean_export(path_to_export, export_type)
}
\arguments{
\item{path_to_export}{path to  export}

\item{export_type}{Specify which bank export will be used.}
}
\description{
Title
}
\keyword{internal}
