% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ledger.R
\name{create_ledger}
\alias{create_ledger}
\title{Title}
\usage{
create_ledger(path_to_ledgerdir, export_type, path_to_export)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files.}

\item{export_type}{Specify which bank export will be used.}

\item{path_to_export}{Path to CSV export file.}
}
\description{
Title
}
