% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_short_ledger.R
\name{create_short_ledger}
\alias{create_short_ledger}
\title{Write short ledger to disk using DKB CSV export}
\usage{
create_short_ledger(path_to_export)
}
\arguments{
\item{path_to_export}{Path to DKB CSV export file}
}
\description{
Wrapper around format_export(), which writes short ledger to disc
}
