% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wide_ledger.R
\name{create_wide_ledger}
\alias{create_wide_ledger}
\title{Write wide ledger to disc}
\usage{
create_wide_ledger(path_to_ledgerdir)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files.}
}
\description{
Reads short ledger and mapping table in working directory,
joins them and writes a CSV.
}
