% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dkbexport.R
\name{read_dkbexport}
\alias{read_dkbexport}
\title{Write short ledger to disk using DKB CSV export}
\usage{
read_dkbexport(path_to_export)
}
\arguments{
\item{path_to_export}{Path to DKB CSV export file}
}
\description{
Wrapper around format_export(), which writes short ledger to disc
}
