% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_balance.R
\name{setup_balance}
\alias{setup_balance}
\title{Add initial balance entry to ledger}
\usage{
setup_balance(path_to_export, export_type, path_to_ledgerdir)
}
\arguments{
\item{path_to_export}{Banking CSV export}

\item{path_to_ledgerdir}{Path to directory that contains all ledger files.}
}
\description{
Add initial balance entry to ledger
}
\keyword{internal}
