% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_balance_bbb.R
\name{setup_balance_bbb}
\alias{setup_balance_bbb}
\title{Title}
\usage{
setup_balance_bbb(path_to_export, ledger)
}
\arguments{
\item{path_to_export}{path_to_export}

\item{ledger}{ledger}
}
\description{
Title
}
