% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_ledger.R
\name{sort_ledger}
\alias{sort_ledger}
\title{Sort ledger}
\usage{
sort_ledger(ledger, ledger_type)
}
\arguments{
\item{ledger}{Dataframe containing ledger}

\item{ledger_type}{"short" or "wide"}
}
\description{
Sort ledger
}
\keyword{internal}
