% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ledger_dir.R
\name{test_ledger_dir}
\alias{test_ledger_dir}
\title{Test for ledgerdir.}
\usage{
test_ledger_dir(path_to_ledgerdir)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files.}
}
\description{
Test for ledgerdir.
}
\keyword{internal}
