% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_maptab.R
\name{test_maptab}
\alias{test_maptab}
\title{Tests for mapping table.}
\usage{
test_maptab(path_to_maptab)
}
\arguments{
\item{path_to_maptab}{Path to mapping table.}
}
\description{
Tests for mapping table.
}
\keyword{internal}
