% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_short_ledger.R
\name{test_short_ledger}
\alias{test_short_ledger}
\title{Tests for short ledger.}
\usage{
test_short_ledger(path_to_shortledger)
}
\arguments{
\item{path_to_shortledger}{Path to short ledger.}
}
\description{
Tests for short ledger.
}
\keyword{internal}
