% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_export.R
\name{transform_export}
\alias{transform_export}
\title{Title}
\usage{
transform_export(path_to_export, export_type)
}
\arguments{
\item{path_to_export}{path to  export}

\item{export_type}{Specify which bank export will be used.}
}
\description{
Transforms the export CSV dataframe to the short ledger format.
The final columns are:

* date
* amount
* recipient
* date_custom
* year
* quarter
* month
* recipient_clean_custom
* amount_custom
* type
* label1_custom
* label2_custom
* label3_custom
}
\keyword{internal}
