% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_export_dkb.R
\name{transform_export_dkb}
\alias{transform_export_dkb}
\title{Transformation for DKB bank}
\usage{
transform_export_dkb(path_to_export)
}
\arguments{
\item{export}{}
}
\description{
Transformation for DKB bank
}
