% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_balance.R
\name{update_balance}
\alias{update_balance}
\title{Update balance column in ledger}
\usage{
update_balance(path_to_ledgerdir)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files.}
}
\description{
Update balance column in ledger
}
\keyword{internal}
