% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpate_cdates.R
\name{update_cdates}
\alias{update_cdates}
\title{Update Custom Dates}
\usage{
update_cdates(path_to_ledgerdir)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files.}
}
\description{
Recalculates year, quarter and month columns using date column
(default) or date_custom column (user input)
}
