% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ledger.R
\name{update_ledger}
\alias{update_ledger}
\title{Wrapper for updating mapping table, balance and maps}
\usage{
update_ledger(path_to_ledgerdir)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files}
}
\description{
Wrapper for updating mapping table, balance and maps
}
