% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_maptab.R
\name{update_maptab}
\alias{update_maptab}
\title{Update the mapping table}
\usage{
update_maptab(path_to_ledgerdir)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files}
}
\description{
Update the mapping table
}
\keyword{internal}
