% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_short_ledger.R
\name{update_short_ledger}
\alias{update_short_ledger}
\title{Wrapper for updating cdates, maptab and balance}
\usage{
update_short_ledger(path_to_ledgerdir)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files.}
}
\description{
Wrapper for updating cdates, maptab and balance
}
