% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_initbalance.R
\name{write_initbalance}
\alias{write_initbalance}
\title{Write initial balance to disk}
\usage{
write_initbalance(dkb_export)
}
\arguments{
\item{dkb_export}{DKB CSV export}
}
\description{
Reads the DKB CSV export and writes an initial entry into the
working directory. It contains the ledger starting date as well the initial balance.
This is used as a starting point for a cumulative history plot.
}
