% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_short_ledger.R
\name{append_short_ledger}
\alias{append_short_ledger}
\title{Appends short ledgers}
\usage{
append_short_ledger(path_to_export)
}
\arguments{
\item{path_to_export}{path to dkb export}
}
\description{
Creates a new short ledger and appends it to the existing one.
To avoid overlapping date ranges, the report discards the most recent days
entries in the existing ledger. The new ledger will start at that cut off
date.
}
