% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dkb.R
\name{clean_dkb}
\alias{clean_dkb}
\title{Title}
\usage{
clean_dkb(dkbexport_df)
}
\arguments{
\item{dkbexport_df}{DKV CSV export as dataframe}
}
\value{
Input Dataframe cleaned to contain date, recipient and amount columns
}
\description{
Title
}
