% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ledger.R
\name{create_ledger}
\alias{create_ledger}
\title{Creates the standard ledger}
\usage{
create_ledger(path_to_ledgerdir, path_to_export, export_type)
}
\arguments{
\item{path_to_ledgerdir}{Path to directory that contains all ledger files.}

\item{path_to_export}{Path to CSV export file.}

\item{export_type}{Specify which bank export will be used.}
}
\description{
Creates the standard ledger
}
