% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wide_ledger.R
\name{create_wide_ledger}
\alias{create_wide_ledger}
\title{Write wide ledger to disc}
\usage{
create_wide_ledger()
}
\description{
Reads short ledger and mapping table in working directory,
joins them and writes a CSV.
}
