% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_export.R
\name{format_export}
\alias{format_export}
\title{Transform cleaned export to short ledger format}
\usage{
format_export(export_clean)
}
\arguments{
\item{export_clean}{CSV dataframe.}
}
\value{
Ledger in short format.
}
\description{
Transforms the DKV CSV dataframe to the short ledger format.
The final columns are:
\itemize{
\item date
\item amount
\item recipient
\item date_custom
\item year
\item quarter
\item month
\item recipient_clean_custom
\item amount_custom
\item type
\item label1_custom
\item label2_custom
\item label3_custom
}
}
\keyword{internal}
