% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_ledger.R
\name{sort_ledger}
\alias{sort_ledger}
\title{Sort ledger columns}
\usage{
sort_ledger(path_to_ledgerdir)
}
\arguments{
\item{ledger}{Dataframe containing ledger}
}
\description{
Sort ledger columns
}
\keyword{internal}
