% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_export.R
\name{transform_export}
\alias{transform_export}
\title{Transforms export to ledger format}
\usage{
transform_export(path_to_export, export_type)
}
\arguments{
\item{path_to_export}{Path to export}

\item{export_type}{Specify which bank export will be used}
}
\description{
Transforms export to ledger format
}
\keyword{internal}
