% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_export_bbb.R
\name{transform_export_bbb}
\alias{transform_export_bbb}
\title{Transformation for BBB bank}
\usage{
transform_export_bbb(path_to_export)
}
\arguments{
\item{export}{}
}
\description{
Transformation for BBB bank
}
