% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpate_cdates.R
\name{udpate_cdates}
\alias{udpate_cdates}
\title{Update Custom Dates}
\usage{
udpate_cdates()
}
\description{
Recalculates year, quarter and month columns using date column
(default) or date_custom column (user input)
}
