% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_initbalance.R
\name{write_initbalance}
\alias{write_initbalance}
\title{Add initial balance entry to short ledger.}
\usage{
write_initbalance(path_to_export, export_type, path_to_ledgerdir)
}
\arguments{
\item{path_to_export}{Banking CSV export}

\item{path_to_ledgerdir}{Path to directory that contains all ledger files.}
}
\description{
Reads the banking CSV export and writes an initial entry into the
working directory. It contains the ledger starting date as well the initial balance.
This is used as a starting point for a cumulative history plot.
}
\keyword{internal}
